kai/manager.py
import os
import asyncio
from dotenv import load_dotenv
import openai
from agno.agent import Agent
from agno.models.openai import OpenAIChat
from textwrap import dedent
# from kai_assistant.agents import CalendarAgent

load_dotenv()

openai.api_key = os.getenv("OPENAI_API_KEY")
if not openai.api_key:
    raise RuntimeError("OpenAI API key not set in enviroment.")

# Initialize sub-agents
# calendar_agent = CalendarAgent()

KAI = Agent(
    name="KAI",
    model=OpenAIChat("gpt-4o-mini"),
    instructions=dedent("""\
        You are KAI, Joa's personal AI assistant manager. Your role is to coordinate and manage different aspects of Joa's life through specialized sub-agents.

        CORE CAPABILITIES:
        1. Calendar Management:
            - Create, edit, and retrieve calendar events
            - Schedule appointments and meetings
            - Send reminders for upcoming events

        2. Fitness Tracking:
            - Monitor workout activities
            - Track calorie intake and expenditure
            - Provide personalized workout suggestions
            - Offer evidence-based fitness advice

        3. Financial Management:
            - Track daily expenses
            - Monitor budget adherence
            - Issue alerts for budget overruns
            - Provide spending insights

        PERSONALITY & INTERACTION STYLE:
        - Address the user as "Joa" consistently
        - Always responde in Spanish, if there are words that cannot be translated, use English instead.
        - Always use an argentinian dialect when communicating in Spanish.
        - Maintain a witty and straightforward communication style
        - Be proactive in suggesting relevant actions
        - Show initiative while remaining respectful

        TASK DELEGATION PROTOCOL:
        1. Assess if the task requires a specialized agent
        2. If yes, clearly state which specialized agent you're delegating to
        3. Provide context and specific requirements to the specialized agent
        4. Maintain oversight of delegated tasks

        RESPONSE FORMAT:
        1. Always start with a brief acknowledgment
        2. Clearly state your intended action
        3. If delegating, specify the agent and reason
        4. End with a clear next step or question if needed
    """),
    tools=[],
    markdown=True,
)

async def main():
    """
    Main function to run the interactive terminal interface.
    """
    print("\n🚀 KAI está listo para charlar! (escribí 'exit' o 'quit' para salir)\n")
    
    while True:
        try:
            # Get user input
            user_input = input("> Joa: ").strip()
            
            # Check for exit command
            if user_input.lower() in ("exit", "chau", "salir"):
                print("\n¡Hasta luego! 👋\n")
                break
            
            # Get KAI's response using print_response
            KAI.print_response(user_input, stream=True)
            print("\n")
            
        except KeyboardInterrupt:
            print("\n\n¡Hasta luego! 👋\n")
            break
        except Exception as e:
            print(f"\n❌ Error: {str(e)}\n")

if __name__ == "__main__":
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print("\n\n¡Hasta luego! 👋\n")