pyproject.toml
[project]
name = "kai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.4.2",
    "dedent>=0.5",
    "google-api-python-client>=2.168.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "openai>=1.76.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "setuptools>=80.0.0",
    "twilio>=9.5.2",
]
